// src/app/write/types.ts
export interface EditArticle {
  onEdit: true;
  idArticleEdit: string | undefined;
  titleEdit: string | undefined;
  valueEdit: string | undefined;
  thumbnailEdit: string | undefined;
  categoryEdit: string | undefined;
  visibilityEdit: string | undefined;
  komentarFieldEdit: string | undefined;
}